refactor: 하드코딩 키값 상수화, 공통 로직 분리, fix: statetag 처리 로직 버그 수정, 앉기 공격 버그 수정

+++ Engine/Util/Delegate.h
@@ -10,13 +10,13 @@ struct DelegateHandle
     {
     }
 
-    DelegateHandle(int id) : handle(id)
-    { 
+    DelegateHandle(unsigned long long id) : handle(id)
+    {
     }
-    
+
     bool IsValid() const { return handle > 0;}
-    
-    bool operator==(const DelegateHandle& other) const 
+
+    bool operator==(const DelegateHandle& other) const
     {
         return (handle == other.handle);
     }
